ci(travis): update autotests for windows

+++ autotest/t999_test.py
@@ -1,6 +1,5 @@
 # Test the download_and_unzip functionality of pymake
 
-import sys
 import os
 import shutil
 import pymake
@@ -28,57 +27,42 @@ def which(program):
     return None
 
 
-def repo_latest_assets(github_repo):
-    """
-    Return a dictionary containing the file name and the link to the asset
-    contained in a github repository.
+def test_latest_version():
+    mfexes_repo_name = 'MODFLOW-USGS/executables'
+    version = pymake.repo_latest_version(mfexes_repo_name)
+    test_version = '3.0'
+    msg = 'returned version ({}) '.format(version) + \
+          'is not equal to defined version ({})'.format(test_version)
+    assert version == test_version, msg
+    return
 
-    Parameters
-    ----------
-    github_repo : str
-        Repository name, such as MODFLOW-USGS/modflow6
 
-    Returns
-    -------
-    result_dict : dict
-        dictionary of file names and links
-
-    """
-    import requests
-    import json
-    repo_url = 'https://api.github.com/repos/{}'.format(github_repo)
-
-    assets = None
-    request_url = '{}/releases/latest'.format(repo_url)
-    print('Requesting from: {}'.format(request_url))
-    r = requests.get(request_url)
-    if (r.ok):
-        jsonobj = json.loads(r.text or r.content)
-        assets = jsonobj['assets']
-    else:
-        assert assets, 'Could not find latest executables from ' + request_url
-
-    result_dict = {}
-    for asset in assets:
-        k = asset['name']
-        v = asset['browser_download_url']
-        result_dict[k] = v
-    return result_dict
+def test_assets():
+    mfexes_repo_name = 'MODFLOW-USGS/executables'
+    assets = pymake.repo_latest_assets(mfexes_repo_name)
+    keys = assets.keys()
+    test_keys = ['mac.zip', 'linux.zip', 'win32.zip', 'win64.zip']
+    for key in keys:
+        msg = 'unknown key ({}) found in github repo assets'.format(key)
+        assert key in test_keys, msg
+    return
 
 
 def test_download_and_unzip():
     pth = './temp/t999'
     pymake.getmfexes(pth, '3.0')
     for f in os.listdir(pth):
-        fname = os.path.join(pth, f)
-        errmsg = '{} not executable'.format(fname)
-        assert which(fname) is not None, errmsg
+        fpth = os.path.join(pth, f)
+        if not os.path.isdir(fpth):
+            errmsg = '{} not executable'.format(fpth)
+            assert which(fpth) is not None, errmsg
 
     # clean up exe's
     for f in os.listdir(pth):
         fpth = os.path.join(pth, f)
-        print('Removing ' + f)
-        os.remove(fpth)
+        if not os.path.isdir(fpth):
+            print('Removing ' + f)
+            os.remove(fpth)
 
     # clean up directory
     if os.path.isdir(pth):
@@ -89,4 +73,6 @@ def test_download_and_unzip():
 
 
 if __name__ == '__main__':
+    test_latest_version()
+    test_assets()
     test_download_and_unzip()
